Map and Camera Movement
Added the black and white map and set the panel to open on specific spots on SEECS and NBS.

Added the ability to move the camera around. There is also a script that lets you zoom in and out using the mouse wheel, but I've removed it from the main camera for now because the screen bugs out when you open the panel by clicking a department.